refactor(interface-cli): extract small helpers (config loader, confirm, interface filter)
Pull three small clusters out of `main.rs` into a new
`crates/interface-cli/src/cli_helpers.rs`:

- `CliConfirmation` — interactive confirmation prompt impl of
  `ConfirmationCallback` (~20 lines)
- `ConfigLoadMessage` + `load_config_messages` — read+parse
  `assistant.toml` with deferred-logging messages (~40 lines)
- `normalize_worker_interface`, `parse_interface_selection`,
  `interface_selected` — worker-interface name normalisation and
  comma-separated filter parsing (~30 lines)

main.rs shrinks 1693 → 1605 lines. cargo test -p assistant-cli reports
the same 59 lib tests passing; make lint is clean.
